+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Join storybook
 - Indicator component
+- Indicator storybook
 
 ## [0.7.4] - 2025-02-15